feat: improve agent-server interaction and DQ checks registration

+++ agent/backend/src/main/resources/db/migration/V1.0__create_initial_schema.sql
@@ -27,7 +27,7 @@ CREATE TABLE result
     check_name       VARCHAR(255),
     check_id         INTEGER,
     raw_value        INTEGER,
-    obfuscated_value INTEGER,
+    obfuscated_value REAL,
     report_id        INTEGER,
     warning_threshold INTEGER,
     error_threshold INTEGER,
